+++ src/icons/charging.png
