+++ src/icons/discharging.png
