+++ src/icons/low-battery.png
