+++ src/icons/not-charging.png
